chore: align with actual API.
