+++ src/common/Model.ts
@@ -11,6 +11,7 @@ import {
     LocRequest,
     LocRequestStatus,
 } from './types/ModelTypes';
+import Identity from './types/Identity';
 
 export interface FetchRequestSpecification {
     legalOfficerAddress?: string,
@@ -118,10 +119,7 @@ export interface CreateLocRequest {
     ownerAddress?: string;
     requesterAddress?: string;
     description?: string;
-    firstName?: string;
-    lastName?: string;
-    email?: string;
-    phone?: string;
+    userIdentity?: Identity;
 }
 
 export async function createLocRequest(
