+++ src/wallet-user/transaction-protection/RequestedLocs.tsx
@@ -1,5 +1,6 @@
 import React, { useCallback, useEffect, useState } from 'react';
 import { useForm } from 'react-hook-form';
+import { Option } from '@polkadot/types';
 
 import { useCommonContext } from '../../common/CommonContext';
 import Table, { Cell, EmptyTableMessage, DateCell } from '../../common/Table';
@@ -7,8 +8,22 @@ import LocStatusCell from '../../common/LocStatusCell';
 import Button from '../../common/Button';
 import Dialog from '../../common/Dialog';
 import { CreateLocRequest, createLocRequest } from '../../common/Model';
+import UserIdentity from '../../common/types/Identity';
+import { useUserContext } from '../UserContext';
+import { RecoveryConfig } from '../../logion-chain/Recovery';
 
 import LocCreationForm, { FormValues } from './LocCreationForm';
+
+function shouldShowIdentityFields(
+    legalOfficer: string | null,
+    recoveryConfig: Option<RecoveryConfig> | null
+): boolean {
+    if(legalOfficer === null || recoveryConfig === null || recoveryConfig.isNone) {
+        return true;
+    } else {
+        return !recoveryConfig.unwrap().friends.toArray().map(accountId => accountId.toString()).includes(legalOfficer);
+    }
+}
 
 export default function RequestedLocs() {
     const { colorTheme, pendingLocRequests, accounts, refresh, axios } = useCommonContext();
@@ -24,25 +39,37 @@ export default function RequestedLocs() {
         }
     });
     const [ selectedLegalOfficer, setSelectedLegalOfficer ] = useState<string | null>(null);
+    const { recoveryConfig } = useUserContext();
+
+    const showIdentityFields = shouldShowIdentityFields(selectedLegalOfficer, recoveryConfig);
 
     const submit = useCallback((formValues: FormValues) => {
+        let userIdentity: UserIdentity | undefined;
+        if(showIdentityFields) {
+            userIdentity = {
+                firstName: formValues.firstName || "",
+                lastName: formValues.lastName || "",
+                email: formValues.email || "",
+                phoneNumber: formValues.phone || "",
+            };
+        } else {
+            userIdentity = undefined;
+        }
+
         (async function() {
             const currentAddress = accounts!.current!.address;
             const request: CreateLocRequest = {
                 ownerAddress: formValues.legalOfficer,
                 requesterAddress: currentAddress,
                 description: formValues.description,
-                firstName: formValues.firstName,
-                lastName: formValues.lastName,
-                email: formValues.email,
-                phone: formValues.phone,
+                userIdentity,
             }
             await createLocRequest!(axios!, request);
             reset();
             setRequestLoc(false);
             refresh!();
         })();
-    }, [ axios, accounts, setRequestLoc, refresh, reset ]);
+    }, [ axios, accounts, setRequestLoc, refresh, reset, showIdentityFields ]);
 
     useEffect(() => {
         const subscription = watch(({ legalOfficer }) => setSelectedLegalOfficer(legalOfficer));
@@ -106,6 +133,7 @@ export default function RequestedLocs() {
                     errors={ errors }
                     colors={ colorTheme.dialog }
                     legalOfficer={ selectedLegalOfficer }
+                    showIdentityFields={ showIdentityFields }
                 />
             </Dialog>
         </>
